code refactor - moved utils folder to common
