+++ src/client/components/patientPortal/dashboard/dates/index.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { Link } from "react-router-dom";
-import { parseDate } from "../../../../utils/util";
+import { parseDate } from "../../../../services/common/util";
 
 class Dates extends Component {
 	constructor(props) {
