+++ src/client/services/clinicPortalServices/orderSearchService.js
@@ -1,6 +1,6 @@
 //import { NormalModuleReplacementPlugin } from "webpack";
 import { serviceConstants } from "../common/constants";
-import { getUserAuthToken } from "../../utils/util";
+import { getUserAuthToken } from "../common/util";
 
 export const fetchOrderMasterData = (facilityId) => {
 	//debugger;
