+++ src/client/services/clinicPortalServices/patientEditService.js
@@ -1,5 +1,6 @@
 import { serviceConstants } from "../common/constants";
-import { getUserAuthToken } from "../../utils/util";
+import { getUserAuthToken } from "../common/util";
+
 export const updatePatientData = (editParams) => {
 	//console.log(editParams);
 	var token  = getUserAuthToken();
