+++ src/client/services/patientSignupServices/patientSignupFormService.js
@@ -1,6 +1,6 @@
 
 import { serviceConstants } from "../common/constants";
-import { getUserAuthToken } from "../../utils/util";
+import { getUserAuthToken } from "../common/util";
 
 export const patientSignup = (patientDetails) => {
     //console.log("patientSignup",patientDetails);
